requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.32.3
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
